Fix fast-mode lineup positions and extract AITacticsService
Fast mode used to land goalkeepers at fullback and forwards at centre-back
when several preferred starters were tired or unavailable: the per-team
selection bucketed players by the coarse `position_group` (CB/LB/RB all
"Defender") and fell back to "fill remaining slots with the highest-OVR
player regardless of position" when no same-group sub existed. The 11 it
produced then went into `FormationRecommender::fillByForceAssignment`,
which placed whatever was left into whatever slot was empty. Replace the
selection with FormationRecommender manual pins keyed on each preferred
player's natural slot (LW preferred over LM, etc.) and let the
recommender's own position-aware passes fill the rest; tired starters
are dropped from the pin set only when a rested compat-100 substitute
exists for their slot.

Then extract `AITacticsService` to own the AI tactical brain — formation
choice, mentality, instructions, team-average, and opponent prediction.
LineupService no longer owns these and `predictOpponentTactics` is no
longer a half-duplicate of `ensureTeamLineup`'s AI-team branch; both
surfaces call the same service. ShowLineup and ShowOpponentAnalysis
pre-load the opponent squad and pass it in, removing the
`matchDate`/`competitionId` plumbing that only existed for the lookup.
