+++ tests/Unit/AIMentalityBiasTest.php
@@ -7,14 +7,14 @@ use App\Modules\Lineup\Enums\DefensiveLineHeight;
 use App\Modules\Lineup\Enums\Mentality;
 use App\Modules\Lineup\Enums\PlayingStyle;
 use App\Modules\Lineup\Enums\PressingIntensity;
+use App\Modules\Lineup\Services\AITacticsService;
 use App\Modules\Lineup\Services\FormationBiasResolver;
 use App\Modules\Lineup\Services\FormationRecommender;
-use App\Modules\Lineup\Services\LineupService;
 use PHPUnit\Framework\TestCase;
 
 class AIMentalityBiasTest extends TestCase
 {
-    private LineupService $lineupService;
+    private AITacticsService $aiTactics;
 
     protected function setUp(): void
     {
@@ -22,10 +22,10 @@ class AIMentalityBiasTest extends TestCase
         // selectAIMentality and selectAIInstructions are pure functions of
         // their inputs; they do not touch the injected collaborators, so
         // raw instances are fine for this unit slice.
-        $this->lineupService = new LineupService(
+        $this->aiTactics = new AITacticsService(
             new FormationRecommender(),
-            $this->createMock(CalendarService::class),
             new FormationBiasResolver(),
+            $this->createMock(CalendarService::class),
         );
     }
 
@@ -33,25 +33,25 @@ class AIMentalityBiasTest extends TestCase
     {
         // Elite team at home vs similar-strength opponent → BALANCED was the
         // documented baseline before bias was introduced.
-        $mentality = $this->lineupService->selectAIMentality('elite', true, 80, 80, 0);
+        $mentality = $this->aiTactics->selectAIMentality('elite', true, 80, 80, 0);
         $this->assertSame(Mentality::BALANCED, $mentality);
     }
 
     public function test_negative_bias_pushes_baseline_toward_defensive(): void
     {
         // -2 should ladder-shift BALANCED two steps down → DEFENSIVE.
-        $mentality = $this->lineupService->selectAIMentality('elite', true, 80, 80, -2);
+        $mentality = $this->aiTactics->selectAIMentality('elite', true, 80, 80, -2);
         $this->assertSame(Mentality::DEFENSIVE, $mentality);
     }
 
     public function test_positive_bias_pushes_baseline_toward_attacking(): void
     {
         // +1 shifts a 'mid'-tier away vs similar (DEFENSIVE) → BALANCED.
-        $balanced = $this->lineupService->selectAIMentality('continental', false, 75, 75, 1);
+        $balanced = $this->aiTactics->selectAIMentality('continental', false, 75, 75, 1);
         $this->assertSame(Mentality::BALANCED, $balanced);
 
         // +2 from the same baseline lands on ATTACKING (clamped at top).
-        $attacking = $this->lineupService->selectAIMentality('continental', false, 75, 75, 2);
+        $attacking = $this->aiTactics->selectAIMentality('continental', false, 75, 75, 2);
         $this->assertSame(Mentality::ATTACKING, $attacking);
     }
 
@@ -59,11 +59,11 @@ class AIMentalityBiasTest extends TestCase
     {
         // Already DEFENSIVE — shifting further down stays DEFENSIVE rather
         // than walking off the ladder.
-        $stillDefensive = $this->lineupService->selectAIMentality('local', false, 60, 75, -2);
+        $stillDefensive = $this->aiTactics->selectAIMentality('local', false, 60, 75, -2);
         $this->assertSame(Mentality::DEFENSIVE, $stillDefensive);
 
         // Already ATTACKING — extra positive bias caps at ATTACKING.
-        $stillAttacking = $this->lineupService->selectAIMentality('elite', true, 85, 75, 2);
+        $stillAttacking = $this->aiTactics->selectAIMentality('elite', true, 85, 75, 2);
         $this->assertSame(Mentality::ATTACKING, $stillAttacking);
     }
 
@@ -71,7 +71,7 @@ class AIMentalityBiasTest extends TestCase
     {
         // The early-return guards inputs we cannot reason about — bias must
         // not push a guarded "we don't know" into a confident extreme.
-        $result = $this->lineupService->selectAIMentality(null, true, 75, 75, 2);
+        $result = $this->aiTactics->selectAIMentality(null, true, 75, 75, 2);
         $this->assertSame(Mentality::BALANCED, $result);
     }
 
@@ -79,7 +79,7 @@ class AIMentalityBiasTest extends TestCase
     {
         // Mid-tier team at home vs similar (baseline BALANCED / STANDARD /
         // NORMAL). +2 bias should walk every dimension up the ladder.
-        [$style, $pressing, $defLine] = $this->lineupService->selectAIInstructions(
+        [$style, $pressing, $defLine] = $this->aiTactics->selectAIInstructions(
             'continental',
             true,
             75,
@@ -96,7 +96,7 @@ class AIMentalityBiasTest extends TestCase
     {
         // Same baseline as above, but -2 bias — every dimension to the
         // defensive end of its ladder.
-        [$style, $pressing, $defLine] = $this->lineupService->selectAIInstructions(
+        [$style, $pressing, $defLine] = $this->aiTactics->selectAIInstructions(
             'continental',
             true,
             75,
